rename lbfgs.c and lbfgs_nocuda.c
