+++ src/lib/Dirac/Makefile
@@ -16,14 +16,14 @@ LIBPATH=
 GLIBI=-I/usr/include/glib-2.0 -I/usr/lib/glib-2.0/include -I/usr/lib/x86_64-linux-gnu/glib-2.0/include/ -I/usr/lib64/glib-2.0/include
 GLIBL=-lglib-2.0
 
-OBJECTS= lmfit_nocuda.o clmfit.o lbfgs_nocuda.o myblas.o robustlm.o updatenu.o robust_lbfgs_nocuda.o robust_batchmode_lbfgs_nocuda.o rtr_solve.o  rtr_solve_robust.o manifold_average.o consensus_poly.o rtr_solve_robust_admm.o admm_solve.o mdl.o
+OBJECTS= lmfit_nocuda.o clmfit.o lbfgs.o myblas.o robustlm.o updatenu.o robust_lbfgs_nocuda.o robust_batchmode_lbfgs_nocuda.o rtr_solve.o  rtr_solve_robust.o manifold_average.o consensus_poly.o rtr_solve_robust_admm.o admm_solve.o mdl.o
 
 default:libdirac.a
 lmfit_nocuda.o:lmfit_nocuda.c
 	$(CC) $(CFLAGS) $(INCLUDES) $(GLIBI)  -c $<
 clmfit.o:clmfit.c
 	$(CC) $(CFLAGS) $(INCLUDES) $(GLIBI)  -c $<
-lbfgs_nocuda.o:lbfgs_nocuda.c
+lbfgs.o:lbfgs.c
 	$(CC) $(CFLAGS) $(INCLUDES) $(GLIBI)  -c $<
 myblas.o:myblas.c
 	$(CC) $(CFLAGS) $(INCLUDES) $(GLIBI)  -c $<
