+++ src/lib/Dirac/Makefile.gpu
@@ -21,13 +21,13 @@ INCLUDES=-I.
 LIBPATH= $(CUDALIB)
 
 
-OBJECTS=lmfit.o lbfgs.o myblas.o mderiv.o clmfit_cuda.o clmfit.o barrier.o robust.o robustlm.o oslmfit.o mderiv_fl.o clmfit_fl.o updatenu.o robust_fl.o manifold_fl.o rtr_solve_cuda.o rtr_solve_robust_cuda.o manifold_average.o consensus_poly.o rtr_solve_robust_cuda_admm.o rtr_solve_robust_admm.o admm_solve.o load_balance.o mdl.o 
+OBJECTS=lmfit.o lbfgs_cuda.o myblas.o mderiv.o clmfit_cuda.o clmfit.o barrier.o robust.o robustlm.o oslmfit.o mderiv_fl.o clmfit_fl.o updatenu.o robust_fl.o manifold_fl.o rtr_solve_cuda.o rtr_solve_robust_cuda.o manifold_average.o consensus_poly.o rtr_solve_robust_cuda_admm.o rtr_solve_robust_admm.o admm_solve.o load_balance.o mdl.o 
 
 
 default:libdirac.a
 lmfit.o:lmfit.c
 	$(CC) $(CFLAGS) $(INCLUDES) $(GLIBI)  -c $<
-lbfgs.o:lbfgs.c
+lbfgs_cuda.o:lbfgs_cuda.c
 	$(CC) $(CFLAGS) $(INCLUDES) $(GLIBI)  -c $<
 myblas.o:myblas.c
 	$(CC) $(CFLAGS) $(INCLUDES) $(GLIBI) -c $<
