+++ src/lib/Dirac/lbfgs.c
@@ -19,201 +19,16 @@
 
 
 #include "Dirac.h"
-#include <cuda.h>
-#include <cuda_runtime_api.h>
-#include <cuda_runtime.h>
 #include <pthread.h>
-#include <math.h>
-
-
-static void
-checkCudaError(cudaError_t err, char *file, int line)
-{
-#ifdef CUDA_DEBUG
-    if(!err)
-        return;
-    fprintf(stderr,"GPU (CUDA): %s %s %d\n", cudaGetErrorString(err),file,line);
-    exit(EXIT_FAILURE);
-#endif
-}
-/************************ pipeline **************************/
-/* data struct shared by all threads */
-typedef struct gb_data_b_ {
-  int status[2]; /* 0: do nothing, 
-              1: allocate GPU  memory, attach GPU
-              2: free GPU memory, detach GPU 
-              3,4..: do work on GPU 
-              99: reset GPU memory (memest all memory) */
-  thread_gpu_data *lmdata[2]; /* two for each thread */
-
-  /* GPU related info */
-  cublasHandle_t cbhandle[2]; /* CUBLAS handles */
-  cusolverDnHandle_t solver_handle[2]; /* solver handles */
-  double *gWORK[2]; /* GPU buffers */
-  int64_t data_size[2]; /* size of buffer (bytes), size gradient vector has different lengths, will be different for each thread */
-  /* different pointers to GPU data */
-  double *cxo[2]; /* data vector */
-  double *ccoh[2]; /* coherency vector */
-  double  *cpp[2]; /* parameter vector */
-  double *cgrad[2]; /* gradient vector */
-  short *cbb[2]; /* baseline map */
-  int *cptoclus[2]; /* param to cluster map */
-
-  /* for cost calculation */
-  int Nbase[2];
-  int boff[2];
-  double fcost[2];
-
-  /* for robust LBFGS */
-  int do_robust;
-} gbdata_b;
-
-/* slave thread 2GPU function */
-static void *
-pipeline_slave_code_b(void *data)
-{
- cudaError_t err;
-
- slave_tdata *td=(slave_tdata*)data;
- gbdata_b *dp=(gbdata_b*)(td->pline->data);
- int tid=td->tid;
- int Nbase=(dp->lmdata[tid]->Nbase)*(dp->lmdata[tid]->tilesz);
- int M=dp->lmdata[tid]->M;
- int N=dp->lmdata[tid]->N;
- int Nparam=(dp->lmdata[tid]->g_end-dp->lmdata[tid]->g_start+1);
- int m=dp->lmdata[tid]->m;
-
- while(1) {
-  sync_barrier(&(td->pline->gate1)); /* stop at gate 1*/
-  if(td->pline->terminate) break; /* if flag is set, break loop */
-  sync_barrier(&(td->pline->gate2)); /* stop at gate 2 */
-  /* do work */
-  if (dp->status[tid]==PT_DO_CDERIV) {
-    /* copy the current solution to device */
-    err=cudaMemcpy(dp->cpp[tid], dp->lmdata[tid]->p, m*sizeof(double), cudaMemcpyHostToDevice);
-    checkCudaError(err,__FILE__,__LINE__);
-    if (!dp->do_robust) {
-     cudakernel_lbfgs_r(dp->lmdata[tid]->ThreadsPerBlock, dp->lmdata[tid]->BlocksPerGrid, Nbase, dp->lmdata[tid]->tilesz, M, N, Nparam, dp->lmdata[tid]->g_start, dp->cxo[tid], dp->ccoh[tid], dp->cpp[tid], dp->cbb[tid], dp->cptoclus[tid], dp->cgrad[tid]);
-    } else {
-     cudakernel_lbfgs_r_robust(dp->lmdata[tid]->ThreadsPerBlock, dp->lmdata[tid]->BlocksPerGrid, Nbase, dp->lmdata[tid]->tilesz, M, N, Nparam, dp->lmdata[tid]->g_start, dp->cxo[tid], dp->ccoh[tid], dp->cpp[tid], dp->cbb[tid], dp->cptoclus[tid], dp->cgrad[tid],dp->lmdata[tid]->robust_nu);
-    }
-    /* read back the result */
-    err=cudaMemcpy(&(dp->lmdata[tid]->g[dp->lmdata[tid]->g_start]), dp->cgrad[tid], Nparam*sizeof(double), cudaMemcpyDeviceToHost);
-    checkCudaError(err,__FILE__,__LINE__);
-  } else if (dp->status[tid]==PT_DO_CCOST) {
-   /* divide total baselines by 2 */
-   int BlocksPerGrid=(dp->Nbase[tid]+dp->lmdata[tid]->ThreadsPerBlock-1)/dp->lmdata[tid]->ThreadsPerBlock;
-   int  boff=dp->boff[tid];
-   /* copy the current solution to device */
-   err=cudaMemcpy(dp->cpp[tid], dp->lmdata[tid]->p, m*sizeof(double), cudaMemcpyHostToDevice);
-   checkCudaError(err,__FILE__,__LINE__);
-   if (!dp->do_robust) {
-    dp->fcost[tid]=cudakernel_lbfgs_cost(dp->lmdata[tid]->ThreadsPerBlock, BlocksPerGrid, dp->Nbase[tid], boff, M, N, Nbase, &dp->cxo[tid][8*boff], &dp->ccoh[tid][boff*8*M], dp->cpp[tid], &dp->cbb[tid][boff*2], dp->cptoclus[tid]); 
-   } else {
-    dp->fcost[tid]=cudakernel_lbfgs_cost_robust(dp->lmdata[tid]->ThreadsPerBlock, BlocksPerGrid, dp->Nbase[tid], boff, M, N, Nbase, &dp->cxo[tid][8*boff], &dp->ccoh[tid][boff*8*M], dp->cpp[tid], &dp->cbb[tid][boff*2], dp->cptoclus[tid], dp->lmdata[tid]->robust_nu);
-   }
-  } else if (dp->status[tid]==PT_DO_AGPU) {
-    attach_gpu_to_thread1(select_work_gpu(MAX_GPU_ID,td->pline->thst),&dp->cbhandle[tid],&dp->solver_handle[tid],&dp->gWORK[tid],dp->data_size[tid]);
-    err=cudaMalloc((void**)&(dp->cxo[tid]),dp->lmdata[tid]->n*sizeof(double));
-    checkCudaError(err,__FILE__,__LINE__);
-    err=cudaMalloc((void**)&(dp->ccoh[tid]),Nbase*8*M*sizeof(double));
-    checkCudaError(err,__FILE__,__LINE__);
-    err=cudaMalloc((void**)&(dp->cpp[tid]),m*sizeof(double));
-    checkCudaError(err,__FILE__,__LINE__);
-    err=cudaMalloc((void**)&(dp->cgrad[tid]),Nparam*sizeof(double));
-    checkCudaError(err,__FILE__,__LINE__);
-    err=cudaMalloc((void**)&(dp->cptoclus[tid]),M*2*sizeof(int));
-    checkCudaError(err,__FILE__,__LINE__);
-    err=cudaMalloc((void**)&(dp->cbb[tid]),Nbase*2*sizeof(short));
-    checkCudaError(err,__FILE__,__LINE__);
-    err=cudaMemcpy(dp->cxo[tid], dp->lmdata[tid]->xo, dp->lmdata[tid]->n*sizeof(double), cudaMemcpyHostToDevice);
-    checkCudaError(err,__FILE__,__LINE__);
-    err=cudaMemcpy(dp->ccoh[tid], dp->lmdata[tid]->coh, Nbase*8*M*sizeof(double), cudaMemcpyHostToDevice);
-    checkCudaError(err,__FILE__,__LINE__);
-    err=cudaMemcpy(dp->cptoclus[tid], dp->lmdata[tid]->ptoclus, M*2*sizeof(int), cudaMemcpyHostToDevice);
-    checkCudaError(err,__FILE__,__LINE__);
-    err=cudaMemcpy(dp->cbb[tid], dp->lmdata[tid]->hbb, Nbase*2*sizeof(short), cudaMemcpyHostToDevice);
-    checkCudaError(err,__FILE__,__LINE__);
-
-  } else if (dp->status[tid]==PT_DO_DGPU) {
-    cudaFree(dp->cxo[tid]);
-    cudaFree(dp->ccoh[tid]);
-    cudaFree(dp->cptoclus[tid]);
-    cudaFree(dp->cbb[tid]);
-    cudaFree(dp->cpp[tid]);
-    cudaFree(dp->cgrad[tid]);
-
-    detach_gpu_from_thread1(dp->cbhandle[tid],dp->solver_handle[tid],dp->gWORK[tid]);
-  }
-
- }
- return NULL;
-}
-
-/* initialize the pipeline
-  and start the slaves rolling */
-static void
-init_pipeline_b(th_pipeline *pline,
-     void *data)
-{
- slave_tdata *t0,*t1;
- pthread_attr_init(&(pline->attr));
- pthread_attr_setdetachstate(&(pline->attr),PTHREAD_CREATE_JOINABLE);
-
- init_th_barrier(&(pline->gate1),3); /* 3 threads, including master */
- init_th_barrier(&(pline->gate2),3); /* 3 threads, including master */
- pline->terminate=0;
- pline->data=data; /* data should have pointers to t1 and t2 */
-
- if ((t0=(slave_tdata*)malloc(sizeof(slave_tdata)))==0) {
-    fprintf(stderr,"no free memory\n");
-    exit(1);
- }
- if ((t1=(slave_tdata*)malloc(sizeof(slave_tdata)))==0) {
-    fprintf(stderr,"no free memory\n");
-    exit(1);
- }
- if ((pline->thst=(taskhist*)malloc(sizeof(taskhist)))==0) {
-    fprintf(stderr,"no free memory\n");
-    exit(1);
- }
-
- init_task_hist(pline->thst);
- t0->pline=t1->pline=pline;
- t0->tid=0;
- t1->tid=1; /* link back t1, t2 to data so they could be freed */
- pline->sd0=t0;
- pline->sd1=t1;
- pthread_create(&(pline->slave0),&(pline->attr),pipeline_slave_code_b,(void*)t0);
- pthread_create(&(pline->slave1),&(pline->attr),pipeline_slave_code_b,(void*)t1);
-}
-
-
-/* destroy the pipeline */
-/* need to kill the slaves first */
-static void
-destroy_pipeline_b(th_pipeline *pline)
-{
- pline->terminate=1;
- sync_barrier(&(pline->gate1));
- pthread_join(pline->slave0,NULL);
- pthread_join(pline->slave1,NULL);
- destroy_th_barrier(&(pline->gate1));
- destroy_th_barrier(&(pline->gate2));
- pthread_attr_destroy(&(pline->attr));
- destroy_task_hist(pline->thst);
- free(pline->thst);
- free(pline->sd0);
- free(pline->sd1);
- pline->data=NULL;
-}
-/************************ end pipeline **************************/
 
 /* use algorithm 9.1 to compute pk=Hk gk */
 /* pk,gk: size m x 1
    s, y: size mM x 1 
    rho: size M x 1 
    ii: true location of the k th values in s,y */
+#ifdef USE_MIC
+__attribute__ ((target(MIC)))
+#endif
 static void
 mult_hessian(int m, double *pk, double *gk, double *s, double *y, double *rho, int M, int ii) {
  int ci;
@@ -222,11 +37,15 @@ mult_hessian(int m, double *pk, double *gk, double *s, double *y, double *rho, i
  double gamma,beta;
 
  if ((alphai=(double*)calloc((size_t)M,sizeof(double)))==0) {
+#ifndef USE_MIC
      fprintf(stderr,"%s: %d: no free memory\n",__FILE__,__LINE__);
+#endif
      exit(1);
  }
  if ((idx=(int*)calloc((size_t)M,sizeof(int)))==0) {
+#ifndef USE_MIC
      fprintf(stderr,"%s: %d: no free memory\n",__FILE__,__LINE__);
+#endif
      exit(1);
  }
  if (M>0) {
@@ -286,20 +105,18 @@ mult_hessian(int m, double *pk, double *gk, double *s, double *y, double *rho, i
 
 /* cubic interpolation in interval [a,b] (a>b is possible)
    to find step that minimizes cost function */
-/*
+/* func: scalar function
    xk: parameter values size m x 1 (at which step is calculated)
    pk: step direction size m x 1 (x(k+1)=x(k)+alphak * pk)
    a/b:  interval for interpolation
-   x: size n x 1 (storage)
    xp: size m x 1 (storage)
-   xo: observed data size n x 1
-   n: size of vector function
    step: step size for differencing 
    adata:  additional data passed to the function
 */
 static double 
 cubic_interp(
-   double *xk, double *pk, double a, double b, double *x, double *xp,  double *xo, int m, int n, double step, void *adata,  th_pipeline *tp, gbdata_b *tpg) {
+   double (*func)(double *p, int m, void *adata),
+   double *xk, double *pk, double a, double b, double *xp,  int m, double step, void *adata) {
 
   double f0,f1,f0d,f1d; /* function values and derivatives at a,b */
   double p01,p02,z0,fz0;
@@ -307,69 +124,23 @@ cubic_interp(
 
   my_dcopy(m,xk,1,xp,1); /* xp<=xk */
   my_daxpy(m,pk,a,xp); /* xp<=xp+(a)*pk */
-  sync_barrier(&(tp->gate1));
-  tpg->lmdata[0]->p=tpg->lmdata[1]->p=xp;
-  tpg->status[0]=tpg->status[1]=PT_DO_CCOST;
-  sync_barrier(&(tp->gate2));
-  sync_barrier(&(tp->gate1));
-  tpg->status[0]=tpg->status[1]=PT_DO_NOTHING;
-  f0=tpg->fcost[0]+tpg->fcost[1];
-  sync_barrier(&(tp->gate2));
-
+  f0=func(xp,m,adata);
   /* grad(phi_0): evaluate at -step and +step */
   my_daxpy(m,pk,step,xp); /* xp<=xp+(a+step)*pk */
-  sync_barrier(&(tp->gate1));
-  tpg->status[0]=tpg->status[1]=PT_DO_CCOST;
-  sync_barrier(&(tp->gate2));
-  sync_barrier(&(tp->gate1));
-  tpg->status[0]=tpg->status[1]=PT_DO_NOTHING;
-  p01=tpg->fcost[0]+tpg->fcost[1];
-  sync_barrier(&(tp->gate2));
-
-
+  p01=func(xp,m,adata);
   my_daxpy(m,pk,-2.0*step,xp); /* xp<=xp+(a-step)*pk */
-  sync_barrier(&(tp->gate1));
-  tpg->status[0]=tpg->status[1]=PT_DO_CCOST;
-  sync_barrier(&(tp->gate2));
-  sync_barrier(&(tp->gate1));
-  tpg->status[0]=tpg->status[1]=PT_DO_NOTHING;
-  p02=tpg->fcost[0]+tpg->fcost[1];
-  sync_barrier(&(tp->gate2));
-
-
+  p02=func(xp,m,adata);
   f0d=(p01-p02)/(2.0*step);
 
+  //FIXME my_dcopy(m,xk,1,xp,1); /* not necessary because xp=xk+(a-step)*pk */
+  //FIXME my_daxpy(m,pk,b,xp); /* xp<=xp+(b)*pk */
   my_daxpy(m,pk,-a+step+b,xp); /* xp<=xp+(b)*pk */
-  sync_barrier(&(tp->gate1));
-  tpg->status[0]=tpg->status[1]=PT_DO_CCOST;
-  sync_barrier(&(tp->gate2));
-  sync_barrier(&(tp->gate1));
-  tpg->status[0]=tpg->status[1]=PT_DO_NOTHING;
-  f1=tpg->fcost[0]+tpg->fcost[1];
-  sync_barrier(&(tp->gate2));
-
-
+  f1=func(xp,m,adata);
   /* grad(phi_1): evaluate at -step and +step */
   my_daxpy(m,pk,step,xp); /* xp<=xp+(b+step)*pk */
-  sync_barrier(&(tp->gate1));
-  tpg->status[0]=tpg->status[1]=PT_DO_CCOST;
-  sync_barrier(&(tp->gate2));
-  sync_barrier(&(tp->gate1));
-  tpg->status[0]=tpg->status[1]=PT_DO_NOTHING;
-  p01=tpg->fcost[0]+tpg->fcost[1];
-  sync_barrier(&(tp->gate2));
-
-
+  p01=func(xp,m,adata);
   my_daxpy(m,pk,-2.0*step,xp); /* xp<=xp+(b-step)*pk */
-  sync_barrier(&(tp->gate1));
-  tpg->status[0]=tpg->status[1]=PT_DO_CCOST;
-  sync_barrier(&(tp->gate2));
-  sync_barrier(&(tp->gate1));
-  tpg->status[0]=tpg->status[1]=PT_DO_NOTHING;
-  p02=tpg->fcost[0]+tpg->fcost[1];
-  sync_barrier(&(tp->gate2));
-
-
+  p02=func(xp,m,adata);
   f1d=(p01-p02)/(2.0*step);
 
 
@@ -392,17 +163,11 @@ cubic_interp(
     fz0=f0+f1;
    } else {
     /* evaluate function for this root */
-    my_daxpy(m,pk,-b+step+a+z0*(b-a),xp); /* xp<=xp+(a+z0(b-a))*pk */
-    sync_barrier(&(tp->gate1));
-    tpg->status[0]=tpg->status[1]=PT_DO_CCOST;
-    sync_barrier(&(tp->gate2));
-    sync_barrier(&(tp->gate1));
-    tpg->status[0]=tpg->status[1]=PT_DO_NOTHING;
-    fz0=tpg->fcost[0]+tpg->fcost[1];
-    sync_barrier(&(tp->gate2));
-
+    //FIXME my_dcopy(m,xk,1,xp,1); /* not necessary because xp=xk+(b-step)*pk */
+    //my_daxpy(m,pk,a+z0*(b-a),xp); /* xp<=xp+(z0)*pk */
+    my_daxpy(m,pk,-b+step+a+z0*(b-a),xp); /* xp<=xp+(a+z0*(b-a))*pk */
+    fz0=func(xp,m,adata);
    }
-   //printf("Val=%lf, [%lf,%lf]\n",fz0,f0,f1);
 
    /* now choose between f0,f1,fz0,fz1 */
    if (f0<f1 && f0<fz0) {
@@ -422,30 +187,31 @@ cubic_interp(
     return b;
    }
   }
-
   /* fallback value */
   return (a+b)*0.5;
 }
 
 
+
+
+
 /*************** Fletcher line search **********************************/
 /* zoom function for line search */
-/* 
+/* func: scalar function
    xk: parameter values size m x 1 (at which step is calculated)
    pk: step direction size m x 1 (x(k+1)=x(k)+alphak * pk)
    a/b: bracket interval [a,b] (a>b) is possible
-   x: size n x 1 (storage)
    xp: size m x 1 (storage)
    phi_0: phi(0)
    gphi_0: grad(phi(0))
-   xo: observed data size n x 1
-   n: size of vector function
+   sigma,rho,t1,t2,t3: line search parameters (from Fletcher) 
    step: step size for differencing 
    adata:  additional data passed to the function
 */
 static double 
 linesearch_zoom(
-   double *xk, double *pk, double a, double b, double *x, double *xp,  double phi_0, double gphi_0, double sigma, double rho, double t1, double t2, double t3, double *xo, int m, int n, double step, void *adata,  th_pipeline *tp, gbdata_b *tpg) {
+   double (*func)(double *p, int m, void *adata),
+   double *xk, double *pk, double a, double b, double *xp, double phi_0, double gphi_0, double sigma, double rho, double t1, double t2, double t3, int m, double step, void *adata) {
 
   double alphaj,phi_j,phi_aj;
   double gphi_j,p01,p02,aj,bj;
@@ -459,57 +225,30 @@ linesearch_zoom(
     /* choose alphaj from [a+t2(b-a),b-t3(b-a)] */
     p01=aj+t2*(bj-aj);
     p02=bj-t3*(bj-aj);
-    alphaj=cubic_interp(xk,pk,p01,p02,x,xp,xo,m,n,step,adata,tp,tpg);
+    alphaj=cubic_interp(func,xk,pk,p01,p02,xp,m,step,adata);
     //printf("cubic intep [%lf,%lf]->%lf\n",p01,p02,alphaj);
 
     /* evaluate phi(alphaj) */
     my_dcopy(m,xk,1,xp,1); /* xp<=xk */
     my_daxpy(m,pk,alphaj,xp); /* xp<=xp+(alphaj)*pk */
-    sync_barrier(&(tp->gate1));
-    tpg->lmdata[0]->p=tpg->lmdata[1]->p=xp;
-    tpg->status[0]=tpg->status[1]=PT_DO_CCOST;
-    sync_barrier(&(tp->gate2));
-    sync_barrier(&(tp->gate1));
-    tpg->status[0]=tpg->status[1]=PT_DO_NOTHING;
-    phi_j=tpg->fcost[0]+tpg->fcost[1];
-    sync_barrier(&(tp->gate2));
-
+    phi_j=func(xp,m,adata);
 
     /* evaluate phi(aj) */
+    //FIXME my_dcopy(m,xk,1,xp,1); /* xp<=xk : not necessary because already copied */
+    //FIXME my_daxpy(m,pk,aj,xp); /* xp<=xp+(aj)*pk */
     my_daxpy(m,pk,-alphaj+aj,xp); /* xp<=xp+(aj)*pk */
-    sync_barrier(&(tp->gate1));
-    tpg->status[0]=tpg->status[1]=PT_DO_CCOST;
-    sync_barrier(&(tp->gate2));
-    sync_barrier(&(tp->gate1));
-    tpg->status[0]=tpg->status[1]=PT_DO_NOTHING;
-    phi_aj=tpg->fcost[0]+tpg->fcost[1];
-    sync_barrier(&(tp->gate2));
-
+    phi_aj=func(xp,m,adata);
 
     if ((phi_j>phi_0+rho*alphaj*gphi_0) || phi_j>=phi_aj) {
       bj=alphaj; /* aj unchanged */
     } else {
      /* evaluate grad(alphaj) */
+     //FIXME my_dcopy(m,xk,1,xp,1); /* xp<=xk */
+     //FIXME my_daxpy(m,pk,alphaj+step,xp); /* xp<=xp+(alphaj+step)*pk */
      my_daxpy(m,pk,-aj+alphaj+step,xp); /* xp<=xp+(alphaj+step)*pk */
-     sync_barrier(&(tp->gate1));
-     tpg->status[0]=tpg->status[1]=PT_DO_CCOST;
-     sync_barrier(&(tp->gate2));
-     sync_barrier(&(tp->gate1));
-     tpg->status[0]=tpg->status[1]=PT_DO_NOTHING;
-     p01=tpg->fcost[0]+tpg->fcost[1];
-     sync_barrier(&(tp->gate2));
-
-
+     p01=func(xp,m,adata);
      my_daxpy(m,pk,-2.0*step,xp); /* xp<=xp+(alphaj-step)*pk */
-     sync_barrier(&(tp->gate1));
-     tpg->status[0]=tpg->status[1]=PT_DO_CCOST;
-     sync_barrier(&(tp->gate2));
-     sync_barrier(&(tp->gate1));
-     tpg->status[0]=tpg->status[1]=PT_DO_NOTHING;
-     p02=tpg->fcost[0]+tpg->fcost[1];
-     sync_barrier(&(tp->gate2));
-
-
+     p02=func(xp,m,adata);
      gphi_j=(p01-p02)/(2.0*step);
 
      /* termination due to roundoff/other errors pp. 38, Fletcher */
@@ -539,7 +278,7 @@ linesearch_zoom(
   }
    
 #ifdef DEBUG
-  printf("Found %lf Interval [%lf,%lf]\n",alphak,a,b);
+  printf("Found %g Interval [%lf,%lf]\n",alphak,a,b);
 #endif
   return alphak;
 }
@@ -547,24 +286,25 @@ linesearch_zoom(
  
 
 /* line search */
-/* 
+/* func: scalar function
    xk: parameter values size m x 1 (at which step is calculated)
    pk: step direction size m x 1 (x(k+1)=x(k)+alphak * pk)
    alpha1: initial value for step
    sigma,rho,t1,t2,t3: line search parameters (from Fletcher) 
-   xo: observed data size n x 1
-   n: size of vector function
+   m: size or parameter vector
    step: step size for differencing 
    adata:  additional data passed to the function
 */
-static double 
+double 
 linesearch(
-   double *xk, double *pk, double alpha1, double sigma, double rho, double t1, double t2, double t3, double *xo, int m, int n, double step, void *adata, th_pipeline *tp, gbdata_b *tpg) {
+   double (*func)(double *p, int m, void *adata),
+   double *xk, double *pk, double alpha1, double sigma, double rho, double t1, double t2, double t3, int m, double step, void *adata) {
+ 
  /* phi(alpha)=f(xk+alpha pk)
   for vector function func 
    f(xk) =||func(xk)||^2 */
   
-  double *x,*xp;
+  double *xp;
   double alphai,alphai1;
   double phi_0,phi_alphai,phi_alphai1;
   double p01,p02;
@@ -572,14 +312,10 @@ linesearch(
   double alphak;
 
   double mu;
-  double tol; /* lower limit for minimization, need to be just about min value of cost function */
+  double tol; /* lower limit for minimization */
 
   int ci;
 
-  if ((x=(double*)calloc((size_t)n,sizeof(double)))==0) {
-     fprintf(stderr,"%s: %d: no free memory\n",__FILE__,__LINE__);
-     exit(1);
-  }
   if ((xp=(double*)calloc((size_t)m,sizeof(double)))==0) {
      fprintf(stderr,"%s: %d: no free memory\n",__FILE__,__LINE__);
      exit(1);
@@ -587,77 +323,42 @@ linesearch(
 
   alphak=1.0;
   /* evaluate phi_0 and grad(phi_0) */
-  sync_barrier(&(tp->gate1));
-  tpg->lmdata[0]->p=tpg->lmdata[1]->p=xk;
-  tpg->status[0]=tpg->status[1]=PT_DO_CCOST;
-  sync_barrier(&(tp->gate2));
-  sync_barrier(&(tp->gate1));
-  tpg->status[0]=tpg->status[1]=PT_DO_NOTHING;
-  phi_0=tpg->fcost[0]+tpg->fcost[1];
-  sync_barrier(&(tp->gate2));
-//printf("GPU cost=%lf\n",phi_0);
+  phi_0=func(xk,m,adata);
   /* select tolarance 1/100 of current function value */
-  tol=MIN(0.01*phi_0,1e-6); 
+  tol=MIN(0.01*phi_0,1e-6);
 
   /* grad(phi_0): evaluate at -step and +step */
   my_dcopy(m,xk,1,xp,1); /* xp<=xk */
   my_daxpy(m,pk,step,xp); /* xp<=xp+(0.0+step)*pk */
-
-  sync_barrier(&(tp->gate1));
-  tpg->lmdata[0]->p=tpg->lmdata[1]->p=xp;
-  tpg->status[0]=tpg->status[1]=PT_DO_CCOST;
-  sync_barrier(&(tp->gate2));
-  sync_barrier(&(tp->gate1));
-  p01=tpg->fcost[0]+tpg->fcost[1];
-  tpg->status[0]=tpg->status[1]=PT_DO_NOTHING;
-  sync_barrier(&(tp->gate2));
-
+  p01=func(xp,m,adata);
   my_daxpy(m,pk,-2.0*step,xp); /* xp<=xp+(0.0-step)*pk */
-  sync_barrier(&(tp->gate1));
-  tpg->status[0]=tpg->status[1]=PT_DO_CCOST;
-  sync_barrier(&(tp->gate2));
-  sync_barrier(&(tp->gate1));
-  p02=tpg->fcost[0]+tpg->fcost[1];
-  tpg->status[0]=tpg->status[1]=PT_DO_NOTHING;
-  sync_barrier(&(tp->gate2));
-
-
+  p02=func(xp,m,adata);
   gphi_0=(p01-p02)/(2.0*step);
-
 
   /* estimate for mu */
   /* mu = (tol-phi_0)/(rho gphi_0) */
   mu=(tol-phi_0)/(rho*gphi_0);
 #ifdef DEBUG
-  printf("cost=%lf grad=%lf mu=%lf, alpha1=%lf\n",phi_0,gphi_0,mu,alpha1);
+  printf("deltaphi=%lf, mu=%lf, alpha1=%lf\n",gphi_0,mu,alpha1);
 #endif
   /* catch if not finite (deltaphi=0 or nan) */
   if (!isnormal(mu)) {
-    free(x);
     free(xp);
 #ifdef DEBUG
-    printf("line interval too small\n");
+  printf("line interval too small\n");
 #endif
     return mu;
   }
 
-
   ci=1;
   alphai=alpha1; /* initial value for alpha(i) : check if 0<alphai<=mu */
-  alphai1=0.0; /* FIXME: tune for GPU (defalut is 0.0) */
+  alphai1=0.0;
   phi_alphai1=phi_0;
   while(ci<10) {
    /* evalualte phi(alpha(i))=f(xk+alphai pk) */
    my_dcopy(m,xk,1,xp,1); /* xp<=xk */
    my_daxpy(m,pk,alphai,xp); /* xp<=xp+alphai*pk */
-   sync_barrier(&(tp->gate1));
-   tpg->status[0]=tpg->status[1]=PT_DO_CCOST;
-   sync_barrier(&(tp->gate2));
-   sync_barrier(&(tp->gate1));
-   phi_alphai=tpg->fcost[0]+tpg->fcost[1];
-   tpg->status[0]=tpg->status[1]=PT_DO_NOTHING;
-   sync_barrier(&(tp->gate2));
-
+   phi_alphai=func(xp,m,adata);
 
    if (phi_alphai<tol) {
      alphak=alphai;
@@ -669,7 +370,7 @@ linesearch(
 
    if ((phi_alphai>phi_0+alphai*gphi_0) || (ci>1 && phi_alphai>=phi_alphai1)) {
       /* ai=alphai1, bi=alphai bracket */
-      alphak=linesearch_zoom(xk,pk,alphai1,alphai,x,xp,phi_0,gphi_0,sigma,rho,t1,t2,t3,xo,m,n,step,adata,tp,tpg);
+      alphak=linesearch_zoom(func,xk,pk,alphai1,alphai,xp,phi_0,gphi_0,sigma,rho,t1,t2,t3,m,step,adata);
 #ifdef DEBUG
       printf("Linesearch : Condition 1 met\n");
 #endif
@@ -677,25 +378,13 @@ linesearch(
    } 
 
    /* evaluate grad(phi(alpha(i))) */
+   //FIXME my_dcopy(m,xk,1,xp,1); /* NOT NEEDED here?? xp<=xk */
+   //FIXME my_daxpy(m,pk,alphai+step,xp); /* xp<=xp+(alphai+step)*pk */
+   /* note that xp  already is xk+alphai. pk, so only add the missing term */
    my_daxpy(m,pk,step,xp); /* xp<=xp+(alphai+step)*pk */
-   sync_barrier(&(tp->gate1));
-   tpg->status[0]=tpg->status[1]=PT_DO_CCOST;
-   sync_barrier(&(tp->gate2));
-   sync_barrier(&(tp->gate1));
-   p01=tpg->fcost[0]+tpg->fcost[1];
-   tpg->status[0]=tpg->status[1]=PT_DO_NOTHING;
-   sync_barrier(&(tp->gate2));
-
+   p01=func(xp,m,adata);
    my_daxpy(m,pk,-2.0*step,xp); /* xp<=xp+(alphai-step)*pk */
-   sync_barrier(&(tp->gate1));
-   tpg->status[0]=tpg->status[1]=PT_DO_CCOST;
-   sync_barrier(&(tp->gate2));
-   sync_barrier(&(tp->gate1));
-   p02=tpg->fcost[0]+tpg->fcost[1];
-   tpg->status[0]=tpg->status[1]=PT_DO_NOTHING;
-   sync_barrier(&(tp->gate2));
-
-
+   p02=func(xp,m,adata);
    gphi_i=(p01-p02)/(2.0*step);
 
    if (fabs(gphi_i)<=-sigma*gphi_0) {
@@ -708,7 +397,7 @@ linesearch(
 
    if (gphi_i>=0) {
      /* ai=alphai, bi=alphai1 bracket */
-     alphak=linesearch_zoom(xk,pk,alphai,alphai1,x,xp,phi_0,gphi_0,sigma,rho,t1,t2,t3,xo,m,n,step,adata,tp,tpg);
+     alphak=linesearch_zoom(func,xk,pk,alphai,alphai1,xp,phi_0,gphi_0,sigma,rho,t1,t2,t3,m,step,adata);
 #ifdef DEBUG
      printf("Linesearch : Condition 3 met\n");
 #endif
@@ -724,7 +413,7 @@ linesearch(
      /* choose by interpolation in [2*alphai-alphai1,min(mu,alphai+t1*(alphai-alphai1)] */
      p01=2.0*alphai-alphai1;
      p02=MIN(mu,alphai+t1*(alphai-alphai1));
-     alphai=cubic_interp(xk,pk,p01,p02,x,xp,xo,m,n,step,adata,tp,tpg);
+     alphai=cubic_interp(func,xk,pk,p01,p02,xp,m,step,adata);
      //printf("cubic interp [%lf,%lf]->%lf\n",p01,p02,alphai);
    }
    phi_alphai1=phi_alphai;
@@ -734,38 +423,78 @@ linesearch(
 
 
 
-  free(x);
   free(xp);
 #ifdef DEBUG
-  printf("Step size=%lf\n",alphak);
+  printf("Step size=%g\n",alphak);
 #endif
   return alphak;
 }
 /*************** END Fletcher line search **********************************/
 
 
-/* note M here  is LBFGS memory size */
+/*************** backtracking line search **********************************/
+/* func: cost function
+   xk: parameter values size m x 1 (at which step is calculated)
+   pk: step direction size m x 1 (x(k+1)=x(k)+alphak * pk)
+   gk: gradient vector size m x 1
+   m: size or parameter vector
+   alpha0: initial alpha
+   adata:  additional data passed to the function
+*/
+static double
+linesearch_backtrack(
+   double (*func)(double *p, int m, void *adata),
+   double *xk, double *pk, double *gk, int m, double alpha0, void *adata) {
+
+  /* Armijo condition  f(x+alpha p) <= f(x) + c alpha p^T grad(f(x)) */
+  const double c=1e-4;
+  double alphak=alpha0;
+  double *xk1,fnew,fold,product;
+  if ((xk1=(double*)calloc((size_t)m,sizeof(double)))==0) {
+     fprintf(stderr,"%s: %d: no free memory\n",__FILE__,__LINE__);
+     exit(1);
+  }
+  /* update parameters xk1=xk+alpha_k *pk */
+  my_dcopy(m,xk,1,xk1,1);
+  my_daxpy(m,pk,alphak,xk1);
+
+  fnew=func(xk1,m,adata);
+  fold=func(xk,m,adata); /* add threshold to make iterations stop at some point FIXME: is this correct/needed? */
+  product=c*my_ddot(m,pk,gk);
+  int ci=0;
+  while (ci<15 && fnew>fold+alphak*product) {
+     alphak *=0.5;
+     my_dcopy(m,xk,1,xk1,1);
+     my_daxpy(m,pk,alphak,xk1);
+     fnew=func(xk1,m,adata);
+     ci++;
+  }
+
+  free(xk1);
+  return alphak;
+}
+/*************** END backtracking line search **********************************/
+
+/* full batch (original) version of LBFGS */
 static int
-lbfgs_fit_common(
-   double *p, double *x, int m, int n, int itmax, int M, int gpu_threads, int do_robust, void *adata) {
+lbfgs_fit_fullbatch(
+   double (*cost_func)(double *p, int m, void *adata),
+   void (*grad_func)(double *p, double *g, int m, void *adata),
+   /* adata: user supplied data, 
+   */
+   /* p:mx1 vector, M: memory size */
+   double *p, int m, int itmax, int M, void *adata) { 
 
   double *gk; /* gradients at both k+1 and k iter */
   double *xk1,*xk; /* parameters at k+1 and k iter */
   double *pk; /* step direction H_k * grad(f) */
 
-  double step; /* FIXME tune for GPU, use larger if far away from convergence */
+  double step=1e-6; /* step for interpolation */
   double *y, *s; /* storage for delta(grad) and delta(p) */
   double *rho; /* storage for 1/yk^T*sk */
   int ci,ck,cm;
   double alphak=1.0;
   
-
-  me_data_t *dp=(me_data_t*)adata;
-  short *hbb;
-  int *ptoclus;
-  int Nbase1=dp->Nbase*dp->tilesz;
-
-  thread_gpu_data threaddata[2]; /* 2 for 2 threads/cards */
 
   if ((gk=(double*)calloc((size_t)m,sizeof(double)))==0) {
      fprintf(stderr,"%s: %d: no free memory\n",__FILE__,__LINE__);
@@ -795,113 +524,18 @@ lbfgs_fit_common(
      fprintf(stderr,"%s: %d: no free memory\n",__FILE__,__LINE__);
      exit(1);
   }
+
+
   if ((rho=(double*)calloc((size_t)M,sizeof(double)))==0) {
      fprintf(stderr,"%s: %d: no free memory\n",__FILE__,__LINE__);
      exit(1);
   }
 
-/*********** following are not part of LBFGS, but done here only for GPU use */
-  /* auxilliary arrays for GPU */
-  if ((hbb=(short*)calloc((size_t)(Nbase1*2),sizeof(short)))==0) {
-     fprintf(stderr,"%s: %d: no free memory\n",__FILE__,__LINE__);
-     exit(1);
-  }
-  /* baseline->station mapping */
-  rearrange_baselines(Nbase1, dp->barr, hbb, dp->Nt);
-
-  /* parameter->cluster mapping */ 
-  /* for each cluster: chunk size, start param index */
-  if ((ptoclus=(int*)calloc((size_t)(2*dp->M),sizeof(int)))==0) {
-     fprintf(stderr,"%s: %d: no free memory\n",__FILE__,__LINE__);
-     exit(1);
-  }
-  for(ci=0; ci<dp->M; ci++) {
-   ptoclus[2*ci]=dp->carr[ci].nchunk;
-   ptoclus[2*ci+1]=dp->carr[ci].p[0]; /* so end at p[0]+nchunk*8*N-1 */
-  }
-  dp->hbb=hbb;
-  dp->ptoclus=ptoclus;
-/*****************************************************************************/
-  /* choose 256 threads per block for high occupancy */
-  int ThreadsPerBlock = gpu_threads;
-
-  /* partition parameters, per each parameter, one thread */
-  /* also account for the no of GPUs using */
-  /* parameters per thread (GPU) */
-  int Nparm=(m+2-1)/2;
-  /* find number of blocks */
-  int BlocksPerGrid =(Nparm+ThreadsPerBlock-1)/ThreadsPerBlock;
-  ci=0;
-  int nth;
-  for (nth=0; nth<2; nth++) {
-   threaddata[nth].ThreadsPerBlock=ThreadsPerBlock;
-   threaddata[nth].BlocksPerGrid=BlocksPerGrid;
-   threaddata[nth].card=nth;
-   threaddata[nth].Nbase=dp->Nbase;
-   threaddata[nth].tilesz=dp->tilesz;
-   threaddata[nth].barr=dp->barr;
-   threaddata[nth].M=dp->M;
-   threaddata[nth].N=dp->N;
-   threaddata[nth].coh=dp->coh;
-   threaddata[nth].xo=x;
-   threaddata[nth].p=p;
-   threaddata[nth].g=gk;
-   threaddata[nth].m=m;
-   threaddata[nth].n=n;
-   threaddata[nth].hbb=dp->hbb;
-   threaddata[nth].ptoclus=dp->ptoclus;
-   threaddata[nth].g_start=ci;
-   threaddata[nth].g_end=ci+Nparm-1;
-   if (threaddata[nth].g_end>=m) {
-     threaddata[nth].g_end=m-1;
-   }
-   /* for robust mode */
-   if (do_robust) { 
-    threaddata[nth].robust_nu=dp->robust_nu;
-   }
-   ci=ci+Nparm;
-  }
-  
-  /* pipeline data */
-  th_pipeline tp;
-  gbdata_b tpg; 
-
-  tpg.do_robust=do_robust;
-  /* divide no of baselines */
-  int Nthb0=(Nbase1+2-1)/2;
-  tpg.Nbase[0]=Nthb0;
-  tpg.Nbase[1]=Nbase1-Nthb0;
-  tpg.boff[0]=0;
-  tpg.boff[1]=Nthb0;
-  
-  tpg.lmdata[0]=&threaddata[0];
-  tpg.lmdata[1]=&threaddata[1];
-  /* calculate total size of memory need to be allocated in GPU, in bytes +2 added to align memory */
-  /* note: we do not allocate memory here, use pinned memory for transfer */
-  //tpg.data_size[0]=(n+(dp->Nbase*dp->tilesz)*8*dp->M+m+(tpg.lmdata[0]->g_end-tpg.lmdata[0]->g_start+1)+2)*sizeof(double)+(2*dp->M*sizeof(int))+(2*dp->Nbase*dp->tilesz*sizeof(char));
-  //tpg.data_size[1]=(n+(dp->Nbase*dp->tilesz)*8*dp->M+m+(tpg.lmdata[1]->g_end-tpg.lmdata[1]->g_start+1)+2)*sizeof(double)+(2*dp->M*sizeof(int))+(2*dp->Nbase*dp->tilesz*sizeof(char));
-  tpg.data_size[0]=tpg.data_size[1]=sizeof(float);
-
-  tpg.status[0]=tpg.status[1]=PT_DO_NOTHING;
-  init_pipeline_b(&tp,&tpg);
-  sync_barrier(&(tp.gate1));
-  tpg.status[0]=tpg.status[1]=PT_DO_AGPU;
-  sync_barrier(&(tp.gate2));
-  sync_barrier(&(tp.gate1));
-  tpg.status[0]=tpg.status[1]=PT_DO_NOTHING;
-  sync_barrier(&(tp.gate2));
-
-/*****************************************************************************/
   /* initial value for params xk=p */
   my_dcopy(m,p,1,xk,1);
-  sync_barrier(&(tp.gate1));
-  threaddata[0].p=threaddata[1].p=xk;
-  tpg.status[0]=tpg.status[1]=PT_DO_CDERIV;
-  sync_barrier(&(tp.gate2));
-  sync_barrier(&(tp.gate1));
-  tpg.status[0]=tpg.status[1]=PT_DO_NOTHING;
-  sync_barrier(&(tp.gate2));
-  
+
+  /*  gradient gk=grad(f)_k */
+  grad_func(xk,gk,m,adata);
   double gradnrm=my_dnrm2(m,gk);
   /* if gradient is too small, no need to solve, so stop */
   if (gradnrm<CLM_STOP_THRESH) {
@@ -915,11 +549,13 @@ lbfgs_fit_common(
 #ifdef DEBUG
   printf("||grad||=%g step=%g\n",gradnrm,step);
 #endif
-
-  cm=0;
-  ci=0;
- 
+  
+  cm=0; /* cycle in 0..(M-1)m (in strides of m)*/
+  ci=0; /* cycle in 0..(M-1) */
+  
+  printf("cost=%g\n",cost_func(xk,m,adata));
   while (ck<itmax && isnormal(gradnrm) && gradnrm>CLM_STOP_THRESH) {
+printf("iter %d gradnrm %g\n",ck,gradnrm);
    /* mult with hessian  pk=-H_k*gk */
    if (ck<M) {
     mult_hessian(m,pk,gk,s,y,rho,ck,ci);
@@ -930,13 +566,12 @@ lbfgs_fit_common(
 
    /* linesearch to find step length */
    /* parameters alpha1=10.0,sigma=0.1, rho=0.01, t1=9, t2=0.1, t3=0.5 */
-   /* FIXME: update paramters for GPU gradient */
-   alphak=linesearch(xk,pk,10.0,0.1,0.01,9,0.1,0.5,x,m,n,step,adata,&tp, &tpg);
+   alphak=linesearch(cost_func,xk,pk,10.0,0.1,0.01,9,0.1,0.5,m,step,adata);
+   
    /* check if step size is too small, or nan, then stop */
    if (!isnormal(alphak) || fabs(alphak)<CLM_EPSILON) {
     break;
    }
-
    /* update parameters xk1=xk+alpha_k *pk */
    my_dcopy(m,xk,1,xk1,1);
    my_daxpy(m,pk,alphak,xk1);
@@ -950,15 +585,9 @@ lbfgs_fit_common(
    my_dcopy(m,gk,1,&y[cm],1); 
    my_dscal(m,-1.0,&y[cm]);
 
-   /* update gradient */
-  sync_barrier(&(tp.gate1));
-  tpg.lmdata[0]->p=tpg.lmdata[1]->p=xk1;
-  tpg.status[0]=tpg.status[1]=PT_DO_CDERIV;
-  sync_barrier(&(tp.gate2));
-  sync_barrier(&(tp.gate1));
-  tpg.status[0]=tpg.status[1]=PT_DO_NOTHING;
-  sync_barrier(&(tp.gate2));
 
+   grad_func(xk1,gk,m,adata);
+   gradnrm=my_dnrm2(m,gk);
    /* yk=yk+gk1 */
    my_daxpy(m,gk,1.0,&y[cm]);
 
@@ -978,15 +607,23 @@ lbfgs_fit_common(
    } else {
     cm=ci=0;
    }
+
+#ifdef DEBUG
+  printf("iter %d alpha=%g ||grad||=%g\n",ck,alphak,gradnrm);
+#endif
+
+  printf("cost=%g\n",cost_func(xk,m,adata));
   }
 
 
  /* copy back solution to p */
  my_dcopy(m,xk,1,p,1);
 
- /* for (ci=0; ci<m; ci++) {
-   printf("grad %d=%lf\n",ci,gk[ci]);
-  } */
+#ifdef DEBUG
+ // for (ci=0; ci<m; ci++) {
+ //  printf("grad %d=%lf\n",ci,gk[ci]);
+ // } 
+#endif
 
   free(gk);
   free(xk1);
@@ -995,32 +632,303 @@ lbfgs_fit_common(
   free(s);
   free(y);
   free(rho);
-  free(hbb);
-  free(ptoclus);
-  dp->hbb=NULL;
-  dp->ptoclus=NULL;
-
-  /******** free threads ***************/
-  sync_barrier(&(tp.gate1)); /* sync at gate 1*/
-  tpg.status[0]=tpg.status[1]=PT_DO_DGPU;
-  sync_barrier(&(tp.gate2)); /* sync at gate 2*/
-
-  destroy_pipeline_b(&tp);
-
   return 0;
 }
 
 
+typedef struct thread_data_vecdot_t_ {
+ int start;
+ int length;
+ double *y;
+ double *a;
+ double *b;
+} thread_data_vecdot_t;
 
-/* wrapper functions */
-int
-lbfgs_fit(
-   double *p, double *x, int m, int n, int itmax, int M, int gpu_threads, void *adata) {
-  return lbfgs_fit_common(p, x, m, n, itmax, M, gpu_threads, 0, adata);
+
+static void *
+outer_product_threadfn(void *data) {
+ thread_data_vecdot_t *t=(thread_data_vecdot_t*)data;
+ int ci;
+ for (ci=t->start; ci<t->start+t->length; ci++) {
+  t->y[ci] += t->a[ci]*t->b[ci];
+ }
+ return NULL;
 }
 
+/* y = y + a.* b
+   a,b,y: vectors of size m
+   Nt: no. of threads 
+*/ 
+static int
+parallel_outer_product(int m,double *__restrict y,double *__restrict a,double *__restrict b,int Nt) {
+  int nth,nth1,ci;
+  int Nthb0,Nthb;
+  pthread_attr_t attr;
+  pthread_t *th_array;
+  thread_data_vecdot_t *threaddata;
+
+  /* setup threads */
+  pthread_attr_init(&attr);
+  pthread_attr_setdetachstate(&attr,PTHREAD_CREATE_JOINABLE);
+
+  if ((th_array=(pthread_t*)malloc((size_t)Nt*sizeof(pthread_t)))==0) {
+   fprintf(stderr,"%s: %d: No free memory\n",__FILE__,__LINE__);
+   exit(1);
+  }
+  if ((threaddata=(thread_data_vecdot_t*)malloc((size_t)Nt*sizeof(thread_data_vecdot_t)))==0) {
+    fprintf(stderr,"%s: %d: No free memory\n",__FILE__,__LINE__);
+    exit(1);
+  }
+
+  /* calculate min values a thread can handle */
+  Nthb0=(m+Nt-1)/Nt;
+  /* iterate over threads, allocating indices per thread */
+  ci=0;
+  for (nth=0;  nth<Nt && ci<m; nth++) {
+   if (ci+Nthb0<m) {
+     Nthb=Nthb0;
+    } else {
+     Nthb=m-ci;
+    }
+    threaddata[nth].start=ci;
+    threaddata[nth].length=Nthb;
+    threaddata[nth].y=y;
+    threaddata[nth].a=a;
+    threaddata[nth].b=b;
+    pthread_create(&th_array[nth],&attr,outer_product_threadfn,(void*)(&threaddata[nth]));
+    /* next data set */
+    ci=ci+Nthb;
+  }
+
+  for(nth1=0; nth1<nth; nth1++) {
+   pthread_join(th_array[nth1],NULL);
+  }
+  pthread_attr_destroy(&attr);
+  free(th_array);
+  free(threaddata);
+  return 0;
+}
+
+
+static int
+lbfgs_fit_minibatch(
+   double (*cost_func)(double *p, int m, void *adata),
+   void (*grad_func)(double *p, double *g, int m, void *adata),
+   /* iter: iteration number, adata: user supplied data, 
+   indata: persistant data that need to be kept between batches */
+   /* p:mx1 vector, M: memory size */
+   double *p, int m, int itmax, int M, void *adata, persistent_data_t *indata) { /* note indata!=NULL is assumed implicitly*/
+
+  double *gk; /* gradients at both k+1 and k iter */
+  double *xk1,*xk; /* parameters at k+1 and k iter */
+  double *pk; /* step direction H_k * grad(f) */
+
+  double *g_min_rold, *g_min_rnew; /* temp storage for updating running averages */
+
+  double *y, *s; /* storage for delta(grad) and delta(p) */
+  double *rho; /* storage for 1/yk^T*sk */
+  int ci,ck,cm;
+  double alphak=1.0;
+  double alphabar=1.0;
+  
+
+  if ((gk=(double*)calloc((size_t)m,sizeof(double)))==0) {
+     fprintf(stderr,"%s: %d: no free memory\n",__FILE__,__LINE__);
+     exit(1);
+  }
+  if ((xk1=(double*)calloc((size_t)m,sizeof(double)))==0) {
+     fprintf(stderr,"%s: %d: no free memory\n",__FILE__,__LINE__);
+     exit(1);
+  }
+  if ((xk=(double*)calloc((size_t)m,sizeof(double)))==0) {
+     fprintf(stderr,"%s: %d: no free memory\n",__FILE__,__LINE__);
+     exit(1);
+  }
+
+  if ((pk=(double*)calloc((size_t)m,sizeof(double)))==0) {
+     fprintf(stderr,"%s: %d: no free memory\n",__FILE__,__LINE__);
+     exit(1);
+  }
+
+
+  /* use y,s pairs from the previous run */
+  /* storage size mM x 1*/
+  s=indata->s;
+  y=indata->y;
+  rho=indata->rho;
+  if (!s || !y || !rho) {
+     fprintf(stderr,"%s: %d: storage must be pre allocated befor calling this function.\n",__FILE__,__LINE__);
+     exit(1);
+  }
+
+  /* initial value for params xk=p */
+  my_dcopy(m,p,1,xk,1);
+
+  /*  gradient gk=grad(f)_k */
+  grad_func(xk,gk,m,adata);
+  double gradnrm=my_dnrm2(m,gk);
+  /* if gradient is too small, no need to solve, so stop */
+  if (gradnrm<CLM_STOP_THRESH) {
+   ck=itmax;
+  } else {
+   ck=0;
+  }
+#ifdef DEBUG
+  printf("||grad||=%g\n",gradnrm);
+#endif
+  
+  ci=indata->vacant; /* cycle in 0..(M-1) */
+  cm=m*ci; /* cycle in 0..(M-1)m (in strides of m)*/
+  
+  while (ck<itmax && isnormal(gradnrm) && gradnrm>CLM_STOP_THRESH) {
+printf("iter %d gradnrm %g\n",ck,gradnrm);
+   /* increment global iteration count */
+   indata->niter++;
+   /* detect if we are at first iteration of a new batch */
+   int batch_changed=(indata->niter>1 && ck==0);
+   /* if the batch has changed, update running averages */
+   if (batch_changed) {
+     /* temp vectors : grad-running_avg(old) , grad - running_avg(new) */
+     /* running_avg_new = running_avg_old + (grad-running_avg(old))/niter */
+     if ((g_min_rold=(double*)calloc((size_t)m,sizeof(double)))==0) {
+      fprintf(stderr,"%s: %d: no free memory\n",__FILE__,__LINE__);
+      exit(1);
+     }
+     if ((g_min_rnew=(double*)calloc((size_t)m,sizeof(double)))==0) {
+      fprintf(stderr,"%s: %d: no free memory\n",__FILE__,__LINE__);
+      exit(1);
+     }
+     my_dcopy(m,gk,1,g_min_rold,1); /* g_min_rold <- grad */
+     my_daxpy(m,indata->running_avg,-1.0,g_min_rold); /* g_min_rold <- g_min_rold - running_avg(old) */
+     my_daxpy(m,g_min_rold,1.0/(double)indata->niter,indata->running_avg); /* running_avg <- running_avg + 1/niter . g_min_rold */
+
+     my_dcopy(m,gk,1,g_min_rnew,1);
+     my_daxpy(m,indata->running_avg,-1.0,g_min_rnew); /* g_min_rnew <- g_min_rnew - running_avg(new) */
+
+     /* this loop should be parallelized/vectorized */
+     /*for (it=0; it<m; it++) {
+       indata->running_avg_sq[it] += g_min_rold[it]*g_min_rnew[it];
+     }*/
+     parallel_outer_product(m,indata->running_avg_sq,g_min_rold,g_min_rnew,indata->Nt);
+
+     /* estimate online variance */
+     alphabar=10.0/(1.0+my_dasum(m,indata->running_avg_sq)/((double)(indata->niter-1)*gradnrm)); 
+     printf("iter=%d running_avg %lf gradnrm %lf alpha=%lf\n",indata->niter,my_dasum(m,indata->running_avg_sq),gradnrm,alphabar);
+     free(g_min_rold);
+     free(g_min_rnew);
+   }
+
+   /* mult with hessian  pk=-H_k*gk */
+   if (indata->nfilled<M) {
+    mult_hessian(m,pk,gk,s,y,rho,indata->nfilled,ci);
+   } else {
+    mult_hessian(m,pk,gk,s,y,rho,M,ci);
+   }
+   my_dscal(m,-1.0,pk);
+
+   /* linesearch to find step length */
+   /* Armijo line search */
+   alphak=linesearch_backtrack(cost_func,xk,pk,gk,m,alphabar,adata);
+   /* check if step size is too small, or nan, then stop */
+   if (!isnormal(alphak) || fabs(alphak)<CLM_EPSILON) {
+    break;
+   }
+   /* update parameters xk1=xk+alpha_k *pk */
+   my_dcopy(m,xk,1,xk1,1);
+   my_daxpy(m,pk,alphak,xk1);
+  
+   if (!batch_changed) {
+   /* calculate sk=xk1-xk and yk=gk1-gk */
+   /* sk=xk1 */ 
+   my_dcopy(m,xk1,1,&s[cm],1); 
+   /* sk=sk-xk */
+   my_daxpy(m,xk,-1.0,&s[cm]);
+   /* yk=-gk */ 
+   my_dcopy(m,gk,1,&y[cm],1); 
+   my_dscal(m,-1.0,&y[cm]);
+   }
+
+   grad_func(xk1,gk,m,adata);
+   gradnrm=my_dnrm2(m,gk);
+ 
+   if (!batch_changed) {
+   /* yk=yk+gk1 */
+   my_daxpy(m,gk,1.0,&y[cm]);
+   
+   /* yk = yk + lm0* sk, to create a trust region */
+   double lm0=1e-6;
+   if (gradnrm>1e3*lm0) {
+    my_daxpy(m,&s[cm],lm0,&y[cm]);
+   }
+   
+   /* calculate 1/yk^T*sk */
+   rho[ci]=1.0/my_ddot(m,&y[cm],&s[cm]);
+   }
+
+   /* update xk=xk1 */
+   my_dcopy(m,xk1,1,xk,1); 
+  
+   //printf("iter %d store %d\n",ck,cm);
+   ck++;
+  
+   if (!batch_changed) {
+   indata->nfilled=(indata->nfilled<M?indata->nfilled+1:M);
+   /* increment storage appropriately */
+   if (cm<(M-1)*m) {
+    /* offset of m */
+    cm=cm+m;
+    ci++;
+    indata->vacant++;
+   } else {
+    cm=ci=0;
+    indata->vacant=0;
+   }
+   }
+
+#ifdef DEBUG
+  printf("iter %d alpha=%g ||grad||=%g\n",ck,alphak,gradnrm);
+#endif
+  }
+
+
+ /* copy back solution to p */
+ my_dcopy(m,xk,1,p,1);
+
+#ifdef DEBUG
+//  for (ci=0; ci<m; ci++) {
+//   printf("grad %d=%lf\n",ci,gk[ci]);
+//  } 
+#endif
+
+  free(gk);
+  free(xk1);
+  free(xk);
+  free(pk);
+  return 0;
+}
+
+/* cost function : return a scalar cost, input : p (mx1) parameters, m: no. of params, adata: additional data
+   grad function: return gradient (mx1): input : p (mx1) parameters, g (mx1) gradient vector, m: no. of params, adata: additional data
+*/ 
+/*
+   p: parameters m x 1 (used as initial value, output final value)
+   itmax: max iterations
+   M: BFGS memory size
+   adata: additional data
+*/
 int
-lbfgs_fit_robust_cuda(
-   double *p, double *x, int m, int n, int itmax, int M, int gpu_threads, void *adata) {
-  return lbfgs_fit_common(p, x, m, n, itmax, M, gpu_threads, 1, adata);
+lbfgs_fit(
+   double (*cost_func)(double *p, int m, void *adata),
+   void (*grad_func)(double *p, double *g, int m, void *adata),
+   /* iter: iteration number, adata: user supplied data, 
+   indata: persistant data that need to be kept between batches */
+   /* p:mx1 vector, M: memory size */
+   double *p, int m, int itmax, int M, void *adata, persistent_data_t *indata) { /* indata=NULL for full batch */
+
+  if (!indata) {
+    lbfgs_fit_fullbatch(cost_func,grad_func,p,m,itmax,M,adata);
+  } else {
+    lbfgs_fit_minibatch(cost_func,grad_func,p,m,itmax,M,adata,indata);
+  }
+  return 0;
 }
